5. Project/0.ETL-data-to-Oracle/requirements.txt
cx_Oracle
pandas
numpy
SQLAlchemy
tqdm
python-dotenv